New WatchStatusButton component for #373. Updated MockActionBar and parent components to use ng-content to add watch status button component
