+++ apps/mobile/pushtracker/src/app/modules/home-tab/home-tab.component.html
@@ -3,7 +3,9 @@
   <ScrollView class="home-tab-layout" height="100%">
     <StackLayout horizontalAlignment="center" row="1" col="0">
       // Mock ActionBar
-      <MockActionBar [showBackNav]="false" [showInfoBtn]="true"></MockActionBar>
+      <MockActionBar [showBackNav]="false" [showInfoBtn]="true">
+        <WatchStatusButton *ngIf="user?.data.control_configuration === CONFIGURATIONS.PUSHTRACKER_WITH_SMARTDRIVE"></WatchStatusButton>
+      </MockActionBar>
 
       // Activity Circles
       <GridLayout rows="auto" columns="auto" horizontalAlignment="center" [height]="screenWidth * 0.6"
