+++ apps/mobile/pushtracker/src/app/modules/shared/components/index.ts
@@ -2,6 +2,7 @@ import { ListPickerSheetComponent, TextFieldSheetComponent, SliderSheetComponent
 import { DataBoxComponent } from './data-box';
 import { MaxTextBoxComponent } from './max-text-box';
 import { MockActionbarComponent } from './mock-actionbar';
+import { WatchStatusButtonComponent } from './watch-status-button';
 
 export const SHARED_COMPONENTS: any[] = [
   DataBoxComponent,
@@ -9,11 +10,12 @@ export const SHARED_COMPONENTS: any[] = [
   MockActionbarComponent,
   ListPickerSheetComponent,
   TextFieldSheetComponent,
-  SliderSheetComponent
+  SliderSheetComponent,
+  WatchStatusButtonComponent
 ];
 
 export * from './bottom-sheets';
 export * from './data-box';
 export * from './max-text-box';
 export * from './mock-actionbar';
-
+export * from './watch-status-button';
