+++ apps/mobile/pushtracker/src/app/modules/shared/components/watch-status-button/index.ts
@@ -1,0 +1,1 @@
+export * from './watch-status-button.component';
