+++ apps/mobile/pushtracker/src/app/modules/shared/components/watch-status-button/watch-status-button.component.html
@@ -1,0 +1,4 @@
+<GridLayout order="1" height="44" width="44" borderRadius="22" 
+  (tap)="onWatchTap()" class="watch-status-btn-layout">
+  <Image [src]="icon" stretch="aspectFit" class="watch-status-btn"></Image>
+</GridLayout>
